:recycle: Refactor lambda code to deploy Container lambda
