+++ .github/workflows/bc_check.yml
@@ -1,0 +1,22 @@
+---
+name: Run Bridgecrew
+on:
+  pull_request:
+    branches: [master]
+  workflow_dispatch:
+
+jobs:
+  bridgecrew:
+    name: Check IaC rules
+    runs-on: ubuntu-latest
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - name: Checkout codebase
+        uses: actions/checkout@v3
+
+      - name: Run Bridgecrew 
+        id: Bridgecrew
+        uses: bridgecrewio/bridgecrew-action@master
+        with:
+          api-key: ${{ secrets.BRIDGECREW_API_KEY }}
+          directory: deployment/terraform
