+++ .github/workflows/linter.yml
@@ -1,0 +1,28 @@
+---
+name: Lint Code Base
+on:
+  pull_request:
+    branches: [master]
+  workflow_dispatch:
+
+jobs:
+  build:
+    name: Lint New Commits
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Code
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Lint Code Base
+        uses: github/super-linter@v4
+        env:
+          VALIDATE_ALL_CODEBASE: false
+          DEFAULT_BRANCH: master
+          VALIDATE_BASH: false
+          VALIDATE_JSCPD: false
+          VALIDATE_MARKDOWN: false
+          VALIDATE_PYTHON_ISORT: false
+          VALIDATE_TERRAFORM_TFLINT: false
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
