+++ .github/workflows/terraform.yml
@@ -1,0 +1,23 @@
+name: Terraform Plan
+on:
+  pull_request:
+    branches: [master]
+    
+jobs:
+
+  Terraform-Fmt:
+    runs-on: ubuntu-latest
+    
+    steps:
+    - uses: actions/checkout@v3
+    - uses: hashicorp/setup-terraform@v1.3.2
+      with:
+        terraform_version: 0.14.6
+      
+    - name: Terraform fmt
+      id: fmt
+      run: terraform fmt -recursive
+
+    - uses: stefanzweifel/git-auto-commit-action@v4
+      with:
+        commit_message: 🎨 Apply terraform fmt
