+++ Dockerfile
@@ -1,8 +1,8 @@
-FROM public.ecr.aws/lambda/python:3.8
+FROM public.ecr.aws/lambda/python:3.9
 
-COPY requirements.txt  .
-RUN  pip3 install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
+WORKDIR /code
 
-COPY mappings.py service.py "${LAMBDA_TASK_ROOT}"/
+COPY src/ ${LAMBDA_TASK_ROOT}
+RUN pip3 install -r ${LAMBDA_TASK_ROOT}/requirements.txt --target "${LAMBDA_TASK_ROOT}" --no-cache-dir
 
-CMD [ "service.handler" ]
+CMD ["app.handler"]
