--- Dockerfile-dev
@@ -1,10 +1,0 @@
-FROM python:3.9-slim-buster
-
-WORKDIR /code
-
-COPY requirements.txt dev-requirements.txt ./
-RUN pip3 install -r dev-requirements.txt -r requirements.txt
-
-COPY . .
-
-CMD ["/bin/bash"]
