+++ deployment/terraform/provider.tf
@@ -1,0 +1,14 @@
+provider "aws" {
+  region = var.region
+
+  default_tags {
+    tags = {
+      commit       = var.lambda_version
+      organization = var.organization
+      application  = var.application
+    }
+  }
+}
+
+data "aws_region" "current" {}
+data "aws_caller_identity" "current" {}
