+++ deployment/terraform/terraform.tfvars.sample
@@ -1,0 +1,6 @@
+# Add variables in this format
+
+environment="<dev/qa/prd>"
+organization="<Organization name>"
+security_group_name="<Dataservice api security group name>"
+es_host="<Elasticsearch Domain hostname>"
