--- dev-requirements.txt
@@ -1,3 +1,0 @@
-pytest==5.1.1
-black==22.10.0
-pylint==2.15.5
