+++ scripts/test
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# scripts/test: Run unit tests in container
+
+docker run --rm -it -v "$PWD":/code --workdir /code python:3.9-slim-buster sh -c "pip install -r src/requirements.txt -r requirements.txt && pytest tests/unit; exit"
