+++ src/copy_member_to_members.py
@@ -1,10 +1,11 @@
 """
 To run this script you need an elasticsearch cluster.
+This script is not meant to used by the lambda.
+It is useful for maintenance.
 """
 from os import environ
-
+import app
 import mappings
-from service import build_es_client
 
 PIPELINE_ID = "member"
 
@@ -18,7 +19,7 @@ def main():
     This script copies (reIndex) documents from member index to members index.
     Assumes that the destination index DOES NOT EXIST.
     """
-    es_client = build_es_client(
+    es_client = app.build_es_client(
         environ.get("es_host", "elasticsearch"),
         environ.get("es_port", "9200"),
         environ.get("es_scheme", "http"),
